% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{scanGRangesForJASPARMotifs}
\alias{scanGRangesForJASPARMotifs}
\title{scanGRangesForJASPARMotifs}
\usage{
scanGRangesForJASPARMotifs(
  granges,
  seqs,
  taxon = "4932",
  db = JASPAR2020::JASPAR2020,
  ncores = 1
)
}
\arguments{
\item{granges}{GRanges object}

\item{seqs}{DNAStringSet objec}

\item{taxon}{String}

\item{db}{JASPAR object}

\item{ncores}{numeric}
}
\value{
GRanges input with extra columns. Each column indicates whether a TF
binding site is found in the GRanges object
}
\description{
Scan a set of GRanges for JASPAR motifs
}
\examples{
ATAC_peaks <- readRDS(here::here('Share/Day1/ATAC_peaks.rds'))
granges <- ATAC_peaks[ATAC_peaks$annot != "non-DA"]
seqlevelsStyle(granges) <- "UCSC"
seqs <- Biostrings::getSeq(BSgenome.Scerevisiae.UCSC.sacCer3::BSgenome.Scerevisiae.UCSC.sacCer3)
x <- scanGRangesForJASPARMotifs(granges, seqs)
}
